Added web service as initial argument.
This update lets you provide your own web service, which is pretty much required for this to be useful!

Be sure the web service spits out JSON in the format:

[
 {
  "id":'anyval',
  "value":"(string)",
  "links":[],
 },
 {
  "id":'anyval2',
  "value":"(string)",
  "links":[],
 },
]

The "links[]" array Follows this same format. Feel free to have complex JSON objects for values or add other properties.